Created BoardCreator class, updated Board class
